Move config related code to _config subdir
