+++ esmvalcore/_config/__init__.py
@@ -1,0 +1,22 @@
+from ._config import (
+    get_activity,
+    get_institutes,
+    get_project_config,
+    load_config_developer,
+    read_config_developer_file,
+    read_config_user_file,
+)
+from ._diagnostics import DIAGNOSTICS, TAGS
+from ._logging import configure_logging
+
+__all__ = (
+    'read_config_user_file',
+    'read_config_developer_file',
+    'load_config_developer',
+    'get_project_config',
+    'get_institutes',
+    'get_activity',
+    'DIAGNOSTICS',
+    'TAGS',
+    'configure_logging',
+)
