+++ tests/integration/test_recipe.py
@@ -10,7 +10,7 @@ import yaml
 from PIL import Image
 
 import esmvalcore
-from esmvalcore._diagnostics import TagsManager
+from esmvalcore._config._diagnostics import TagsManager
 from esmvalcore._recipe import TASKSEP, read_recipe_file
 from esmvalcore._recipe_checks import RecipeError
 from esmvalcore._task import DiagnosticTask
@@ -1333,7 +1333,7 @@ def test_diagnostic_task_provenance(
     monkeypatch,
     config_user,
 ):
-    monkeypatch.setattr(esmvalcore._diagnostics, 'TAGS', TAGS)
+    monkeypatch.setattr(esmvalcore._config, 'TAGS', TAGS)
     monkeypatch.setattr(esmvalcore._recipe, 'TAGS', TAGS)
     monkeypatch.setattr(esmvalcore._task, 'TAGS', TAGS)
